Move Conflict dataclass
